Combine property and tenure data for readability
