+++ src/components/Property/PropertyDetails.tsx
@@ -4,8 +4,7 @@ import BackButton from '../Layout/BackButton'
 import { isCurrentTimeOutOfHours } from '@/utils/helpers/completionDateTimes'
 import Link from 'next/link'
 import WarningInfoBox from '../Template/WarningInfoBox'
-import { Tenure } from '../../models/tenure'
-import { Property } from '../../models/property'
+import { Property, Tenure } from '../../models/propertyTenure'
 
 interface PropertyDetailsProps {
   property: Property
