+++ src/components/Property/RepairsFinder/RepairsFinderFormView.tsx
@@ -12,8 +12,7 @@ import { STATUS_AUTHORISATION_PENDING_APPROVAL } from '@/utils/statusCodes'
 import Meta from '../../Meta'
 import router from 'next/router'
 import { formatRequestErrorMessage } from '@/root/src/utils/errorHandling/formatErrorMessage'
-import { Property } from '@/root/src/models/property'
-import { Tenure } from '@/root/src/models/tenure'
+import { Property, Tenure } from '@/root/src/models/propertyTenure'
 import { isOutOfHoursGas } from './helpers'
 import { Priority } from '@/root/src/models/priority'
 import RaiseWorkOrderSuccessView from './RaiseWorkOrderSuccessView'
