+++ src/components/WorkOrder/MobileWorkingWorkOrder.tsx
@@ -29,8 +29,7 @@ import ErrorMessage from '../Errors/ErrorMessage'
 import PhotoViewList from './Photos/PhotoViewList'
 import WarningInfoBox from '../Template/WarningInfoBox'
 import { formatRequestErrorMessage } from '../../utils/errorHandling/formatErrorMessage'
-import { Property } from '../../models/property'
-import { Tenure } from '../../models/tenure'
+import { Property, Tenure } from '../../models/propertyTenure'
 import { WorkOrderAppointmentDetails } from '../../models/workOrderAppointmentDetails'
 
 const VARIATION_PENDING_APPROVAL_STATUS = 'Variation Pending Approval'
