+++ src/components/WorkOrder/WorkOrderDetails.tsx
@@ -8,9 +8,8 @@ import MultiButton from '../Layout/MultiButton'
 import { WORK_ORDER_ACTIONS } from 'src/utils/workOrderActions'
 import { WorkOrder } from '@/models/workOrder'
 import FollowOnFlag from '../Flags/FollowOnFlag'
-import { Tenure } from '../../models/tenure'
 import { CautionaryAlert } from '../../models/cautionaryAlerts'
-import { Property } from '../../models/property'
+import { Property, Tenure } from '../../models/propertyTenure'
 import { WorkOrderAppointmentDetails } from '../../models/workOrderAppointmentDetails'
 
 interface Props {
