+++ src/components/WorkOrder/WorkOrderHeader.tsx
@@ -8,8 +8,8 @@ import { WorkOrder } from '@/models/workOrder'
 import { CLOSED_STATUS_DESCRIPTIONS_FOR_OPERATIVES } from '@/utils/statusCodes'
 import FurtherWorkRequiredFlag from '../Flags/FurtherWorkRequiredFlag'
 import { CautionaryAlert } from '../../models/cautionaryAlerts'
-import { Tenure } from '../../models/tenure'
 import { WorkOrderAppointmentDetails } from '../../models/workOrderAppointmentDetails'
+import { Tenure } from '../../models/propertyTenure'
 
 interface Props {
   propertyReference: string
