--- src/models/propertyResponse.ts
@@ -1,7 +1,0 @@
-import { Property } from './property'
-import { Tenure } from './tenure'
-
-export interface PropertyResponse {
-  property: Property
-  tenure: Tenure
-}
