--- src/models/tenure.ts
@@ -1,6 +1,0 @@
-export type Tenure = {
-  typeCode: string
-  typeDescription: string
-  tenancyAgreementReference: string
-  id: string
-}
